apiJavaFN

+++ src/main/java/segurospp/seguro/Persona/ClienteRepository.java
@@ -4,5 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+
+    // Optional<Cliente> findByNumeroPoliza(String numeroPoliza);
 }
